refactor: merge task implementations
