--- crates/bevy_tasks/src/wasm_task.rs
@@ -1,106 +1,0 @@
-use alloc::boxed::Box;
-use core::{
-    any::Any,
-    future::Future,
-    panic::{AssertUnwindSafe, UnwindSafe},
-    pin::{pin, Pin},
-    task::{Context, Poll},
-};
-
-use bevy_platform::exports::wasm_bindgen_futures;
-
-/// Wraps an asynchronous task, a spawned future.
-///
-/// Tasks are also futures themselves and yield the output of the spawned future.
-#[derive(Debug)]
-pub struct Task<T>(async_channel::Receiver<Result<T, Panic>>);
-
-impl<T: 'static> Task<T> {
-    pub(crate) fn wrap_future(future: impl Future<Output = T> + 'static) -> Self {
-        let (sender, receiver) = async_channel::bounded(1);
-        wasm_bindgen_futures::spawn_local(async move {
-            // Catch any panics that occur when polling the future so they can
-            // be propagated back to the task handle.
-            let value = CatchUnwind(AssertUnwindSafe(future)).await;
-            let _ = sender.send(value);
-        });
-        Self(receiver)
-    }
-
-    /// When building for Wasm, this method has no effect.
-    /// This is only included for feature parity with other platforms.
-    pub fn detach(self) {}
-
-    /// Returns `true` if the task is finished.
-    pub fn is_finished(&self) -> bool {
-        !self.0.is_empty()
-    }
-
-    /// Requests a task to be cancelled and returns a future that suspends until it completes.
-    /// Returns the output of the future if it has already completed.
-    ///
-    /// # Implementation
-    ///
-    /// When building for Wasm, it is not possible to cancel tasks, which means this is the same
-    /// as just awaiting the task. This method is only included for feature parity with other platforms.
-    pub async fn cancel(self) -> Option<T> {
-        match self.0.recv().await {
-            Ok(Ok(value)) => Some(value),
-            Err(_) => None,
-            Ok(Err(panic)) => {
-                // drop this to prevent the panic payload from resuming the panic on drop.
-                // this also leaks the box but I'm not sure how to avoid that
-                core::mem::forget(panic);
-                None
-            }
-        }
-    }
-}
-
-impl<T> Future for Task<T> {
-    type Output = T;
-    fn poll(self: Pin<&mut Self>, cx: &mut Context<'_>) -> Poll<Self::Output> {
-        let recv = pin!(self.0.recv());
-        match recv.poll(cx) {
-            Poll::Ready(Ok(Ok(value))) => Poll::Ready(value),
-            // NOTE: Propagating the panic here sorta has parity with the async_executor behavior.
-            // For those tasks, polling them after a panic returns a `None` which gets `unwrap`ed, so
-            // using `resume_unwind` here is essentially keeping the same behavior while adding more information.
-            Poll::Ready(Ok(Err(_panic))) => crate::cfg::switch! {{
-                crate::cfg::std => {
-                    std::panic::resume_unwind(_panic)
-                }
-                _ => {
-                    unreachable!("catching a panic is only possible with std")
-                }
-            }},
-            Poll::Ready(Err(_)) => panic!("Polled a task after it was cancelled"),
-            Poll::Pending => Poll::Pending,
-        }
-    }
-}
-
-impl<T> Unpin for Task<T> {}
-
-type Panic = Box<dyn Any + Send + 'static>;
-
-#[pin_project::pin_project]
-struct CatchUnwind<F: UnwindSafe>(#[pin] F);
-
-impl<F: Future + UnwindSafe> Future for CatchUnwind<F> {
-    type Output = Result<F::Output, Panic>;
-    fn poll(self: Pin<&mut Self>, cx: &mut Context) -> Poll<Self::Output> {
-        let f = AssertUnwindSafe(|| self.project().0.poll(cx));
-
-        let result = crate::cfg::switch! {{
-            crate::cfg::std => {
-                std::panic::catch_unwind(f)?
-            }
-            _ => {
-                f()
-            }
-        }};
-
-        result.map(Ok)
-    }
-}
